backend/planner_app/views/user_view.py
from rest_framework import viewsets
from planner_app.models import User
from planner_app.serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
