backend/requirements.txt
Django>=4.2,<4.3
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.14,<3.15
djangorestframework_simplejwt>=5.2,<5.3
django-extensions>=3.1,<4.0
